Move InterfaceSubscriptionsIT test back into 1.dev suite
