+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_15.scala
@@ -10,7 +10,6 @@ package object v1_15 {
   def default(timeoutScaleFactor: Double): Vector[LedgerTestSuite] =
     v1_14.default(timeoutScaleFactor) ++ Vector(
       new InterfaceIT,
-      new InterfaceSubscriptionsIT,
     )
 
   def optional(tlsConfig: Option[TlsConfiguration]): Vector[LedgerTestSuite] =
